# openapi/oh.yaml
openapi: 3.0.1
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /operations/rows:
    put:
      tags:
        - Operations
      operationId: updateOperationRow
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationRowDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Operations
      operationId: newOperationRow
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationRowDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRowDTO"
      security:
        - bearerAuth: [ ]
  /opds/{code}:
    put:
      tags:
        - Opds
      operationId: updateOpd
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpdDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpdDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Opds
      operationId: deleteOpd
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
  /opds/rows/{code}:
    put:
      tags:
        - Opds
      operationId: updateOpdWithOperationRow
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpdWithOperationRowDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpdWithOperationRowDTO"
      security:
        - bearerAuth: [ ]
...
  /patients:
    get:
      tags:
        - Patients
      operationId: getPatients
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 80
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagePatientDTO"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Patients
      operationId: newPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
        - bearerAuth: [ ]
...
  /opds:
    post:
      tags:
        - Opds
      operationId: newOpd
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpdDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpdDTO"
      security:
        - bearerAuth: [ ]
  /opds/rows:
    post:
      tags:
        - Opds
      operationId: newOpdWithOperationRow
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpdWithOperationRowDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpdWithOperationRowDTO"
      security:
        - bearerAuth: [ ]
...
  /patientvaccines/progyear/{year}:
    get:
      tags:
        - Patient Vaccines
      operationId: getProgYear
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
      security:
        - bearerAuth: [ ]
...
  /opds/weekly:
    get:
      tags:
        - Opds
      operationId: getOpdToDayOrWeek
      parameters:
        - name: oneWeek
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpdDTO"
      security:
        - bearerAuth: [ ]
  /opds/search:
    get:
      tags:
        - Opds
      operationId: getOpdByDates
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema:
            type: string
        - name: dateTo
          in: query
          required: true
          schema:
            type: string
        - name: diseaseTypeCode
          in: query
          required: false
          schema:
            type: string
        - name: diseaseCode
          in: query
          required: false
          schema:
            type: string
        - name: ageFrom
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: ageTo
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 200
        - name: sex
          in: query
          required: false
          schema:
            type: string
            default: A
        - name: newPatient
          in: query
          required: false
          schema:
            type: string
            default: A
        - name: patientCode
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 80
        - name: paged
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: wardCode
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOpdDTO"
      security:
        - bearerAuth: [ ]
  /opds/patient/{pcode}:
    get:
      tags:
        - Opds
      operationId: getOpdByPatient
      parameters:
        - name: pcode
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpdWithOperationRowDTO"
      security:
        - bearerAuth: [ ]
  /opds/last/{patientCode}:
    get:
      tags:
        - Opds
      operationId: getLastOpd
      parameters:
        - name: patientCode
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpdDTO"
      security:
        - bearerAuth: [ ]
  /opds/check/progyear:
    get:
      tags:
        - Opds
      operationId: isExistOpdNum
      parameters:
        - name: opdNum
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
  /opds/ProgYear/{year}:
    get:
      tags:
        - Opds
      operationId: getProgressiveYear
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    OperationTypeDTO:
      required:
        - code
        - description
      type: object
      properties:
        code:
          maxLength: 2
          type: string
          description: The operation code
        description:
          maxLength: 50
          type: string
          description: The operation type
    OperationDTO:
      required:
        - code
        - description
        - major
        - opeFor
        - type
      type: object
      properties:
        code:
          maxLength: 10
          type: string
          description: The code of operation
        description:
          maxLength: 50
          type: string
          description: The operation description
        type:
          $ref: "#/components/schemas/OperationTypeDTO"
        opeFor:
          type: string
          description: The operation context
          example: opd_admission
          enum:
            - opd_admission
            - admission
            - opd
        major:
          type: integer
          description: The operation major
          format: int32
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
...
    BillDTO:
      required:
        - amount
        - balance
        - date
        - listName
        - patName
        - patientTrue
        - status
        - update
        - user
      type: object
      properties:
        id:
          type: integer
          format: int32
        patient:
          $ref: "#/components/schemas/PatientDTO"
        listId:
          type: integer
          format: int32
        date:
          type: string
          description: Date of bill creation
          example: 2020-03-19T14:58:00.000Z
        update:
          type: string
          description: Date of bill updated
          example: 2020-03-19T14:58:00.000Z
        listName:
          maxLength: 50
          type: string
          description: Price list name
          example: Basic
        patientTrue:
          type: boolean
          description: Is bill belongs to a patient?
          example: true
        patName:
          maxLength: 100
          type: string
          description: Patient name
          example: Mario Rossi
        status:
          type: string
          description: Bill status
          example: O
        amount:
          type: number
          description: Bill Amount
          format: double
          example: 1000
        balance:
          type: number
          description: Bill balance
          format: double
          example: 1500
        user:
          type: string
          description: user name who create the bill
          example: admin
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        list:
          type: boolean
      description: Class representing a bill
...
    OpdDTO:
      required:
        - age
        - newPatient
        - note
        - sex
        - ward
      type: object
      properties:
        code:
          type: integer
          description: The code of the opd
          format: int32
          example: 3
        date:
          type: string
          description: The date of the admission
        nextVisitDate:
          type: string
          description: The next visit date
        patientCode:
          type: integer
          description: The admitted patient code
          format: int32
        age:
          type: integer
          description: The patient age
          format: int32
          example: 18
        sex:
          type: string
          description: The patient sex
          example: M
        patientName:
          type: string
          description: The patient sex
          example: M
        ageType:
          type: string
          description: Age type
          example: "null"
        note:
          maxLength: 65535
          type: string
          description: The admission note
          example: this is out patient
        prog_year:
          type: integer
          description: A progr. in year for each ward
          format: int32
          example: 18
        disease:
          $ref: "#/components/schemas/DiseaseDTO"
        disease2:
          $ref: "#/components/schemas/DiseaseDTO"
        disease3:
          $ref: "#/components/schemas/DiseaseDTO"
        newPatient:
          type: string
          description: New(N) or Reattendance(R) patient
          example: "N"
        referralFrom:
          type: string
          description: Referral from another unit
          example: R
        referralTo:
          type: string
          description: Referral to another unit
          example: R
        userID:
          type: string
          description: User id
        lock:
          type: integer
          format: int32
          readOnly: true
        hashCode:
          type: integer
          format: int32
          readOnly: true
        reason:
          type: string
          description: Reasons for entry
        anamnesis:
          type: string
          description: History of a medical or psychiatric patient
        allergies:
          type: string
          description: Allergies of patient
        therapies:
          type: string
          description: Current therapies
        prescription:
          maxLength: 255
          type: string
          description: Prescription
        ward:
          $ref: "#/components/schemas/WardDTO"
      description: The opd
    OperationRowDTO:
      required:
        - opDate
        - opResult
        - operation
        - prescriber
      type: object
      properties:
        id:
          type: integer
          format: int32
        operation:
          $ref: "#/components/schemas/OperationDTO"
        prescriber:
          maxLength: 150
          type: string
          description: The presciber of the operation
        opResult:
          maxLength: 250
          type: string
          description: The result of the operation
        opDate:
          type: string
          description: Operation registration date
        remarks:
          maxLength: 250
          type: string
          description: The remark of the operation
        admission:
          $ref: "#/components/schemas/AdmissionDTO"
        opd:
          $ref: "#/components/schemas/OpdDTO"
        bill:
          $ref: "#/components/schemas/BillDTO"
        transUnit:
          type: number
          description: The transunit
          format: float
        hashCode:
          type: integer
          format: int32
          readOnly: true
    OpdWithOperationRowDTO:
      type: object
      properties:
        opdDTO:
          $ref: "#/components/schemas/OpdDTO"
        operationRows:
          type: array
          items:
            $ref: "#/components/schemas/OperationRowDTO"
...
    PageInfoDTO:
      type: object
      properties:
        size:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        nbOfElements:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int64
        totalNbOfElements:
          type: integer
          format: int64
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
    PagePatientDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PatientDTO"
        pageInfo:
          $ref: "#/components/schemas/PageInfoDTO"
    PageOpdDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OpdDTO"
        pageInfo:
          $ref: "#/components/schemas/PageInfoDTO"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT